@@ -8,7 +8,7 @@
     "url": "https://github.com/rodekruis/IBF.git",
     "directory": "services/api-service"
   },
-  "main": "index.js",
+  "main": "index.mjs",
   "scripts": {
     "lint": "eslint \"**/*.{ts,js}\"",
     "fix": "npm run lint -- --fix",
@@ -21,12 +21,12 @@
     "prestart:dev:collect-coverage": "npm install --no-save",
     "seed:dev": "ts-node --require=tsconfig-paths/register src/scripts seed-dev",
     "test": "echo 'No automated tests. Run `npm run test:integration:all` to run API-tests locally. Run `npm run test:unit:all` to run unit-tests locally.' && exit 0",
-    "test:unit:all": "jest --runInBand --config=jest.unit.config.js",
+    "test:unit:all": "jest --runInBand --config=jest.unit.config.mjs",
     "test:unit:watch": "npm run test:unit:all -- --watchAll",
     "test:unit:coverage": "npm run test:unit:all -- --coverage --coverageDirectory=coverage/unit",
-    "test:integration:local": "node --env-file-if-exists=../.env --require=tsconfig-paths/register --expose-gc node_modules/.bin/jest --config=jest.integration.config.js --runInBand --detectOpenHandles --logHeapUsage",
-    "test:integration:all": "node --require=tsconfig-paths/register --expose-gc node_modules/.bin/jest --config=jest.integration.config.js --runInBand --detectOpenHandles --logHeapUsage",
-    "test:integration:debug": "node --inspect-brk=0.0.0.0:9231 --require=tsconfig-paths/register --expose-gc node_modules/.bin/jest --config=jest.integration.config.js --runInBand --detectOpenHandles --logHeapUsage",
+    "test:integration:local": "node --env-file-if-exists=../.env --require=tsconfig-paths/register --expose-gc node_modules/.bin/jest --config=jest.integration.config.mjs --runInBand --detectOpenHandles --logHeapUsage",
+    "test:integration:all": "node --require=tsconfig-paths/register --expose-gc node_modules/.bin/jest --config=jest.integration.config.mjs --runInBand --detectOpenHandles --logHeapUsage",
+    "test:integration:debug": "node --inspect-brk=0.0.0.0:9231 --require=tsconfig-paths/register --expose-gc node_modules/.bin/jest --config=jest.integration.config.mjs --runInBand --detectOpenHandles --logHeapUsage",
     "test:integration:watch": "npm run test:integration:all -- --watchAll",
     "coverage:report:unit": "echo 'This is generated with the test:unit:coverage script.' && exit 1",
     "coverage:report:integration": "nyc report --report-dir coverage/integration --reporter=lcov --reporter=text",
@@ -95,6 +95,7 @@
   "devDependencies": {
     "@automock/adapters.nestjs": "^2.1.0",
     "@compodoc/compodoc": "^1.1.32",
+    "@eslint/js": "^9.39.2",
     "@nestjs/cli": "^11.0.14",
     "@swc/cli": "^0.7.9",
     "@swc/core": "^1.15.8",
@@ -106,7 +107,7 @@
     "@types/passport-jwt": "^4.0.1",
     "@typescript-eslint/eslint-plugin": "^8.52.0",
     "@typescript-eslint/parser": "^8.52.0",
-    "eslint": "^8.57.1",
+    "eslint": "^9.39.2",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-custom-rules": "file:./eslint-plugin-custom-rules/",
     "eslint-plugin-eslint-comments": "^3.2.0",
